+++ R/tapajos.R
@@ -4,14 +4,13 @@
 #' area in the Brazilian 
 #' Amazon \url{http://www.icmbio.gov.br/flonatapajos/}
 #' 
-#' @format An object of class 'sfc_POLYGON' within the bbox 
+#' @format An object of class 'sfc_POLYGON' within the bounding box 
 #' xmin: -55.41127 ymin: -4.114584 
 #' xmax: -54.7973 ymax: -2.751706
 #' @examples 
-#' \donttest{
 #' library("sf")
 #' 
-#' data("tapajos")
+#' data("tapajos", package = "chirps")
 #' 
 #' # sample three points within the Tapajos area
 #' set.seed(1234)
@@ -19,10 +18,6 @@
 #' 
 #' # coerce as sf points
 #' tp_point <- st_as_sf(tp_point)
-#' 
-#' # transform into a geojson
-#' tp_geojson <- sf_to_geojson(tp_point)
-#' }
 #' 
 #' @source The data was provided by the Chico Mendes Institute via
 #' \url{https://www.protectedplanet.net/}.